Renamed cmdUtils.go to config.go and moved getScheddsFromCondor into scheddCache.go
